Refactor tests so state tests are together.
It used to be that the state tests that required a websocket were
lumped in with the general websocket tests.
