+++ backend/tests/base_test.py
@@ -45,3 +45,24 @@ class BaseHttpTest(_BaseTestMixin, tornado.testing.AsyncHTTPTestCase):
       The app to use for testing. """
     # We can just use all the default settings.
     return server.make_app({})
+
+class BaseWebSocketTest(BaseHttpTest):
+  """ Convenience class for writing a test that uses a websocket. """
+
+  def setUp(self):
+    super().setUp()
+
+    # WebSocket URL.
+    self._socket_url = "ws://localhost:%d/app_socket" % (self.get_http_port())
+
+  def _connect_and_run(self, connect_callback, receive_callback=None):
+    """ Connects to a websocket, and runs a callback once the connection has
+    been initiated.
+    Args:
+      connect_callback: The callback to run once the socket is connected.
+      receive_message: The callback to run when a message is received. """
+    connected = tornado.websocket.websocket_connect(self._socket_url,
+                    on_message_callback=receive_callback)
+    self.io_loop.add_future(connected, connect_callback)
+
+    self.wait()
